backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pandas>=2.0.0
numpy>=1.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0
openpyxl>=3.1.0
mysql-connector-python>=8.2.0
python-jose[cryptography]>=3.3.0
google-cloud-storage>=2.10.0
pyarrow>=14.0.0
python-dotenv>=1.0.0